more consistent --yaml handling of dicts in lists
Given an example YAML file as follows:

```
---
- a: A
  b: B
```

This YAML file could be written as JSON as follows:

    [{"a":"A","b":"B"}]

Since the order of keys in a dict (object for JSON) does not
matter, this is identical to the following JSON array:

    [{"b":"B","a":"A"}]

This could be represented in YAML as follows:

```
---
- b: B
  a: A
```

The dict keys a and b have the same indentation level, they
are part of the same list item.  Using the "--yaml" option
should result in printing the matched line and any further
indented lines directly following, according to YAML indentation
rules.

Before this commit, matching the line that starts a list item
printed the complete list item.  Matching on any other line
would not.  This was cause by mixing YAML and normal indentation
rules instead of using only YAML indentation rules.

Now only YAML indentation rules are used whith "--yaml".

In addition to changing the "--yaml" behavior, this commit also
adjusts some existing tests to the new behavior, and adds more
tests for "--yaml".