add --separator --omit tests with two input files
When section --omit --separator is used with two input files,
and some trailing lines of the first file are omitted, but
the first line(s) of the second file are printed, then there
should be a separator above the first line of the second file.

This does not work currently.

Add two tests, a failing one to check for this behavior, and
a passing one where there must be no separator between lines
from the two files.