avoid creating variables inside a for loop
Each variable created inside the for loop will be a new one.
Avoid this, unless this is intended.

See <https://github.com/golang/go/wiki/CommonMistakes>.

Go behavior may change in the future to create new variables
for each loop iteration when using something like:

  for _, i := range …

If this happens, I probably want to change all such loops to
manually create used the iteration variable(s) outside the loop.

See:
 - <https://go.dev/doc/faq#closures_and_goroutines>
 - <https://github.com/golang/go/wiki/LoopvarExperiment>

I only did a cursory check if my uses of loop variables and
their addresses are problematic.  I would expect them to be
OK for now, but I am not sure at all.  This seems to be a
subtle trap in the Go language.