avoid "for ... := ..." loop construct
The for loop := construct has been identified as a source of
bugs, and thus Go version 1.22 is planned to create one new
variable per loop iteration when it is used.

That is not how I understand this construct, and it is not
how it works today.  In order to avoid the coming weirdness,
I plan not to use this construct any more.  Instead I create a
new block around the loop, and define the loop variable inside
this block, but outside the loop, to make the loop variable
lifetime explicit.