six new tests to document current behavior
If all non-blank lines start with non-empty indentation,
and there are empty lines (not even blanks) in the input,
then top-level sections the empty line directly preceding
the section (if there is such a line).  This is similar
for enclosing sections, i.e., the enclosing section might
start with an directly preceding empty line.

I was suprised when I first saw this, but it follows from
the algorithm used to determine the start of a new top-
level section.  For the time being, I intend to keep this
behavior.  I might change it in the future, or add an
option to change this.

The tests are intended to prevent me from _accidentally_
changing this behavior.