rename option --yaml to --yaml-seq-indent
While I had intended section --yaml to work well with
YAML data, this goal is not yet accomplished.  The current
implementation is convenient, i.e., one can use --yaml instead
of the more complicated --indent-re=….  Thus I want to keep
the functionality.  But I want to reserve the --yaml option
name for a better, more complete handling of YAML data that
might be implemented in the future.