bug fix for ignored lines with --enclosing
Before, ignored lines were always treated as outside of a
section.  Now, the inside or outside of a section status is
tracked correctly.  This currently only occurs when using the
--enclosing option.

I do not have a test case where this results in wrong output,
but work on a new feature uncovered the problem.  I want to
fix it independently of the new feature.  But I cannot add a
test with the fix.