new option --headers
The --headers option adds "section headers", i.e., the starting
lines of sections enclosing the selected selection.  The selected
section is also modified to start at the "top level", but not all
lines of the selected top level section need to be selected.

The combination of --top-level with --headers is supported, but
the result is identical to using just --top-level.