document & test --top-level overriding --enclosing
When both options --top-level and --enclosing are used together,
the --enclosing option is ignored.  Describe this in the man page.

Also add a test using "--top-level --enclosing".  A test using
"--enclosing --top-level" already exists.

The rationale is that the top level section always includes
the enclosing section.

In a way it could be said that both options are active even
when only --top-level is actually used, but I think it is
easier to think of this as --top-level overriding --enclosing.
At least I tend to come up with this view time and time again,
before realizing that the --top-level selection includes the
--enclosing selection.