Makefile: gccgo: build statically linked binary
With gccgo, the startup time for a statically linked binary
is shorter than that for a dynamically linked binary.  This
results in significantly longer test execution time when
using the gccgo to build a dynamically linked binary than
using gccgo to build a statically linked binary.

The Go Tool builds a statically linked binary by default.
Thus, this commit aligns the results of using either gccgo
or the Go Tool in this regard.