add options to print a file header
With the new option "--file-header", a header giving the
file name is printed before the first output line from
a file.  The file name is enclosed in a prefix and a suffix
which can be customized using the additional new options
"--file-header-prefix" and "file-header-suffix".

This addresses the feature request from GitHub issue #1.

This commit also adjusts the version number to indicate not
yet released changes.